Major reorganization for separation of concerns and clarity
