+++ packages/lib-panoptes-js/src/index.js
@@ -1,10 +1,9 @@
 const { config } = require('./config')
 const panoptes = require('./panoptes')
-const { projects, projectMocks } = require('./resources/projects')
+const projects = require('./resources/projects')
 
 module.exports = {
   config,
   panoptes,
-  projects,
-  projectMocks
+  projects
 }
