+++ packages/lib-panoptes-js/src/resources/projects/index.js
@@ -1,8 +1,17 @@
-const { projects, projectsEndpoint } = require('./projects')
-const mocks = require('./mocks')
+const { create, get, update, del } = require('./rest')
+const { getBySlug, getWithLinkedResources } = require('./commonRequests')
+const { endpoint } = require('./helpers')
+const { mocks } = require('./mocks')
 
-module.exports = {
-  projects,
-  projectsEndpoint,
-  projectMocks: mocks
+const projects = {
+  create,
+  get,
+  update,
+  delete: del,
+  endpoint,
+  getBySlug,
+  getWithLinkedResources,
+  mocks
 }
+
+module.exports = { projects }
