+++ packages/lib-panoptes-js/src/resources/projects/mocks.js
@@ -1,10 +1,12 @@
-const { userMocks } = require('../users')
+const users = require('../users')
 const { buildMockedMediumResource } = require('../media')
+const { buildResponse } = require('../../helpers')
 
-const projectAvatarResource = buildMockedMediumResource('avatar', 'project')
-const projectBackgroundResource = buildMockedMediumResource('background', 'project')
+// Resources
+const projectAvatar = buildMockedMediumResource('avatar', 'project')
+const projectBackground = buildMockedMediumResource('background', 'project')
 
-const projectOneResource = {
+const projectOne = {
   activity: 0,
   available_languages: ['en'],
   beta_approved: false,
@@ -43,7 +45,7 @@ const projectOneResource = {
   workflow_description: ''
 }
 
-const projectTwoResource = {
+const projectTwo = {
   activity: 100000,
   available_languages: ['en'],
   beta_approved: true,
@@ -82,170 +84,101 @@ const projectTwoResource = {
   workflow_description: ''
 }
 
-const newProjectResponse = {
-  linked: {},
-  links: {},
-  meta: {},
-  projects: [projectOneResource]
-}
-
-const getProjectsResponse = {
-  linked: {},
-  links: {},
-  meta: {},
-  projects: [projectOneResource, projectTwoResource]
-}
-
-const getSingleProjectResponse = {
-  linked: {},
-  links: {},
-  meta: {},
-  projects: [projectTwoResource]
-}
-
-const putProjectResponse = {
-  linked: {},
-  links: {},
-  meta: {},
-  projects: [Object.assign({}, projectTwoResource, { researcher_quote: 'Try my project!' })]
-}
-
-const notFoundResponse = {
-  links: {},
-  meta: {},
-  projects: []
-}
-
-projectRoleOwnerResponse = {
-  links: {},
-  meta: {},
-  project_roles: [
-    { href: "/project_roles/1",
-      id: "1",
-      links: { project: "2", owner: {} },
-      roles: ["owner"]
-    }
-  ]
-}
-
-const projectRolesResponse = {
-  links: {},
-  meta: {},
-  project_roles: [
-    { href: "/project_roles/1",
-      id: "1",
-      links: { project: "2", owner: {} },
-      roles: ["owner"]
-    },
-    { href: "/project_roles/2",
-      id: "2",
-      links: { project: "2", owner: {} },
-      roles: ["expert", "scientist"]
-    }
-  ]
-}
-
-const projectRoleMultipleRolesResponse = {
-  links: {},
-  meta: {},
-  project_roles: [
-    { href: "/project_roles/2",
-      id: "2",
-      links: { project: "2", owner: {} },
-      roles: ["expert", "scientist"]
-    }
-  ]
-}
-
-const projectPageResource = {
-  content: "",
-  created_at: "2016-12-08T20:34:38.300Z",
-  href: "/projects/2/pages/1",
-  id: "1",
-  language: "en",
-  links: { project: "1" },
-  title: "Team",
-  type: "project_pages",
-  updated_at: "2018-05-17T19:09:11.209Z",
-  url_key: "team"
-}
-
-const projectPageResponse = {
-  links: {},
-  meta: {},
-  project_pages: [
-    { content: "",
-      created_at: "2016-12-08T20:34:38.300Z",
-      href: "/projects/2/pages/1",
-      id: "1",
-      language: "en",
-      links: { project: "1" },
-      title: "Team",
-      type: "project_pages",
-      updated_at: "2018-05-17T19:09:11.209Z",
-      url_key: "team"
-    }
-  ]
-}
-
-const projectPagesResponse = {
-  links: {},
-  meta: {},
-  project_pages: [
-    { content: "",
-      created_at: "2016-12-08T20:34:38.300Z",
-      href: "/projects/2/pages/1",
-      id: "1",
-      language: "en",
-      links: { project: "1" },
-      title: "Team",
-      type: "project_pages",
-      updated_at: "2018-05-17T19:09:11.209Z",
-      url_key: "team"
-    },
-    { content: "",
-      created_at: "2016-12-08T20:34:47.085Z",
-      href: "/projects/2/pages/2",
-      id: "2",
-      language: "en",
-      links: { project: "2" },
-      title: "Results",
-      type: "project_pages",
-      updated_at: "2018-05-17T19:16:57.618Z",
-      url_key: "results"
-    }
-  ]
-}
-
-const getProjectResponseWithLinkedResources = {
-  linked: {
-    avatars: [projectAvatarResource],
-    backgrounds: [projectBackgroundResource],
-    owners: userMocks.user,
-    project_pages: [
-      projectPageResource
-    ]
+const projectPages = {
+  team: {
+    content: "",
+    created_at: "2016-12-08T20:34:38.300Z",
+    href: "/projects/2/pages/1",
+    id: "1",
+    language: "en",
+    links: { project: "1" },
+    title: "Team",
+    type: "project_pages",
+    updated_at: "2018-05-17T19:09:11.209Z",
+    url_key: "team"
   },
-  links: {},
-  meta: {},
-  projects: [projectTwoResource]
+  results: {
+    content: "",
+    created_at: "2016-12-08T20:34:47.085Z",
+    href: "/projects/2/pages/2",
+    id: "2",
+    language: "en",
+    links: { project: "2" },
+    title: "Results",
+    type: "project_pages",
+    updated_at: "2018-05-17T19:16:57.618Z",
+    url_key: "results"
+  }
+}
+
+const projectRoles = {
+  owner: {
+    href: "/project_roles/1",
+    id: "1",
+    links: { project: "2", owner: {} },
+    roles: ["owner"]
+  },
+  multipleRoles: {
+    href: "/project_roles/2",
+    id: "2",
+    links: { project: "2", owner: {} },
+    roles: ["expert", "scientist"]
+  }
+}
+
+const resources = {
+  projectAvatar,
+  projectBackground,
+  projectOne,
+  projectTwo,
+  projectPages,
+  projectRoles
+}
+
+// Responses
+
+const createdProject = buildResponse('post', 'projects', [resources.projectOne], {})
+
+const projects = buildResponse('get', 'projects', [resources.projectOne, resources.projectTwo], {})
+
+const project = buildResponse('get', 'projects', [resources.projectTwo], {}) 
+
+const updatedProject = buildResponse('put', 'projects', [resources.projectTwo], {}, { researcher_quote: 'Try my project!' })
+
+const queryNotFound = buildResponse('get', 'projects', [])
+
+const projectRoleOwner = buildResponse('get', 'project_roles', [resources.projectRoles.owner])
+
+const projectRolesResponse = buildResponse('get', 'project_roles', [resources.projectRoles.owner, resources.projectRoles.multipleRoles])
+
+const projectPagesResponse = buildResponse('get', 'project_pages', [resources.projectPages.team, resources, projectPages.results]) 
+
+const projectWithLinkedResources = buildResponse('get', 'projects', [resources.projectTwo], {
+  avatars: [resources.projectAvatar],
+  backgrounds: [resources.projectBackground],
+  owners: users.mocks.resources.user,
+  project_pages: [resources.projectPages]
+})
+
+const responses = {
+  get: {
+    project,
+    projects,
+    queryNotFound,
+    projectRoleOwner,
+    projectRoles: projectRolesResponse,
+    projectPages: projectPagesResponse,
+    projectWithLinkedResources
+  },
+  post: {
+    createdProject
+  },
+  put: {
+    updatedProject
+  }
 }
 
 module.exports = {
-  projectOneResource,
-  projectTwoResource,
-  newProjectResponse,
-  getProjectsResponse,
-  getSingleProjectResponse,
-  putProjectResponse,
-  notFoundResponse,
-  projectPageResource,
-  projectPageResponse,
-  projectPagesResponse,
-  projectRoleOwnerResponse,
-  projectRoleMultipleRolesResponse,
-  projectRolesResponse,
-  projectAvatarResource,
-  projectBackgroundResource,
-  getProjectResponseWithLinkedResources
+  resources,
+  responses
 }
