+++ packages/lib-panoptes-js/src/resources/users/index.js
@@ -1,5 +1,5 @@
 const mocks = require('./mocks')
 
 module.exports = {
-  userMocks: mocks
+  mocks
 }
