+++ packages/lib-panoptes-js/src/resources/users/mocks.js
@@ -13,6 +13,8 @@ const user = {
   valid_email: true
 }
 
+const resources = { user }
+
 module.exports = {
-  user
+  resources
 }
